backend/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
pydantic-settings
tinyfish
apscheduler
groq
voyageai
resend
httpx
python-dotenv
agentops
upstash-redis
python-jose[cryptography]
passlib
websockets
python-multipart
